Textures!
I think the scaling might be off. In a classical implementation, the
world is broken up into a uniform grid, so a full texture can be applied
to each block. Here, there are issues with arbitrarily-sized walls and
textures that may be mismatched in their dimensions.

Still, I'm happy with the performance and the result!